[268444] Update ApplyStereotypeAction to select a resource location https://bugs.eclipse.org/bugs/show_bug.cgi?id=268444
Add a resource selection widget to the Apply Stereotype dialog.  When the user chooses a previously unloaded resource, it may contain stereotype applications, so refresh the dialog accordingly.

Add a listener in the editor that refreshes the model viewer when resources are added/loaded/unloaded to update the presentation of applied stereotypes.